// Akshara/consulting/src/app/clients/shared/client.ts
export class Client {

  id: number;
  typeofClient: string;
  ccDate: Date = new Date();
  companyName: string;
  registrationName: string;
  registrationNumber: string;
  rnumberDocument : File;
  gstNumber: string;
  gstnumberDocument:File;
  rDate: Date = new Date();
  rentalAgrement: File;
  co1Name: string;
  co2Name: string;
  co3Name: string;
  co4Name: string;
  co5Name: string; 
  co1Dob: Date = new Date(); 
  co2Dob: Date = new Date(); 
  co3Dob: Date = new Date(); 
  co4Dob: Date = new Date(); 
  co5Dob: Date = new Date(); 
  pan1Number: string;
  pan2Number: string;
  pan3Number: string;
  pan4Number: string;
  pan5Number: string;
  pan1Doc: File;
  pan2Doc: File;
  pan3Doc: File;
  pan4Doc: File;
  pan5Doc: File;
  aadharNumber1: string;
  aadharNumber2: string;
  aadharNumber3: string;
  aadharNumber4: string;
  aadharNumber5: string;
  aadharNumber1Doc: File;
  aadharNumber2Doc: File;
  aadharNumber3Doc: File;
  aadharNumber4Doc: File;
  aadharNumber5Doc: File;
  contactNumber1: string;
  contactNumber2: string;
  contactNumber3: string;
  contactNumber4: string;
  contactNumber5: string;
  contactEmail1: string;
  contactEmail2: string;
  contactEmail3: string;
  contactEmail4: string;
  contactEmail5: string;
  registeredAddress: string;
  communictionAddress: string;
  bankName: string;
  bankbranchName: string;
  accountNumber: string;
  ifscNumber: string;
  partnershipdeedDoc: File;
  incorporationcertificateDoc: File;
  moaDoc: File;
  aoaDoc: File;
  directorName: string;
  directoridNumber: string;
  directorfatherName: string;
  directorspancardNumber: string;
  directorsaadharNumber: string;
  directorspresentAddress: string;
  directorspermanentAddress: string;
  dateofAppointment: Date = new Date();
  
}